delete composition
